backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0

# LangGraph & AI
langgraph>=0.0.20
langchain>=0.1.0
langchain-anthropic>=0.1.1
langchain-core>=0.1.0
langsmith>=0.1.0
anthropic>=0.17.0
openai>=1.0.0
langchain-openai>=0.0.2
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1

# Data Scraping/APIs
asyncpraw==7.8.1
beautifulsoup4==4.12.3
lxml==5.1.0

# Utilities
python-multipart==0.0.6
websockets==12.0
